aws/AWS-CDK/DynamoDB/create-a-dynamodb-and-grant-access-to-lambda-functions-and-attach-it-to-an-api-gateway/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.23.0
aws-cdk.aws-apigatewayv2-alpha==2.23.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.23.0a0
aws-cdk.aws-batch-alpha==2.23.0a0
aws-cdk.aws-lambda-python-alpha==2.23.0a0
cattrs==22.1.0
constructs==10.1.3
exceptiongroup==1.0.0rc5
flake8==4.0.1
jsii==1.58.0
mccabe==0.6.1
publication==0.0.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.2.0
